feat: チュートリアル再実行導線を About に集約し設定系から削除
チュートリアルは設定項目ではなくヘルプ/案内なので、設定メニューと設定
quickpick (ヘルプ群) の導線を外し、About ウィンドウ上部に独立した CTA
として置く。技術情報やバグ報告と混同しないよう区切って配置。コマンド
パレットの /tutorial は発見導線として維持。
