+++ src/components/deck/DeckSettingsMenu.vue
@@ -2,7 +2,6 @@
 import { computed, ref, toRef } from 'vue'
 
 import { useNativeDialog } from '@/composables/useNativeDialog'
-import { useTutorialStore } from '@/composables/useTutorial'
 import { useVaporTransition } from '@/composables/useVaporTransition'
 import { usePerformanceStore } from '@/stores/performance'
 import { useThemeStore } from '@/stores/theme'
@@ -35,12 +34,6 @@ useNativeDialog(
 )
 
 const windowsStore = useWindowsStore()
-const tutorialStore = useTutorialStore()
-
-function openTutorial(): void {
-  tutorialStore.start()
-  emit('close')
-}
 
 function openToolWindow(
   type:
@@ -75,13 +68,6 @@ function openToolWindow(
       @pointerdown.stop
     >
       <div :class="$style.menuBody">
-        <div :class="$style.categorySection">
-          <button :class="$style.categoryHeader" @click="openTutorial">
-            <i class="ti ti-presentation-analytics" />
-            <span>チュートリアル</span>
-            <i class="ti ti-chevron-right" :class="$style.chevronNav" />
-          </button>
-        </div>
         <div :class="$style.categorySection">
           <button :class="$style.categoryHeader" @click="openToolWindow('appearanceEditor')">
             <i class="ti ti-brush" />
